+++ lib/mattermost/Service.js
@@ -2,7 +2,7 @@
 
 const Co = require('co');
 
-const Api = require('./Api');
+const API = require('./API');
 const Session = require('./Session');
 const User = require('./User');
 const Util = require('../util');
@@ -30,7 +30,7 @@ exports.consume = Co.wrap(function* (res, userInfo) {
   if (!user) {
     const password = User.password();
     const auth = yield Session.getRoot().getAuth();
-    const userObj = yield Api.user.create({ data: marshal(userInfo, password) }, { auth });
+    const userObj = yield API.user.create({ data: marshal(userInfo, password) }, { auth });
     user = yield new User(unmarshal(userObj, password)).insert();
   }
