+++ lib/mattermost/Session.js
@@ -6,7 +6,7 @@ const Co = require('co');
 const Config = require('config');
 const Cookie = require('cookie');
 
-const Api = require('./Api');
+const API = require('./API');
 const Util = require('../util');
 
 // Internal logic
@@ -61,7 +61,7 @@ class Session {
       data = { login_id: credentials.username || credentials.email, password: credentials.password };
     }
 
-    return Api.user.login({ data }, { digest: extractAuth })
+    return API.user.login({ data }, { digest: extractAuth })
       .then((auth) => new Session(credentials, auth));
   }
 
@@ -99,7 +99,7 @@ class Session {
   }
 
   refresh() {
-    return Api.user.login({ data: this.credentials }, { digest: extractAuth })
+    return API.user.login({ data: this.credentials }, { digest: extractAuth })
       .then((auth) => {
         this.update(auth);
         return this;
