+++ lib/mattermost/System.js
@@ -7,7 +7,7 @@ const Co = require('co');
 const Config = require('config');
 const VError = require('verror');
 
-const Api = require('./Api');
+const API = require('./API');
 const Log = require('../../config/logging').get('MATTERMOST_SYSTEM');
 const Session = require('./Session');
 const User = require('./User');
@@ -31,7 +31,7 @@ const getOrCreateCredentials = Co.wrap(function* (credentials, first = true) {
   }
 
   const password = User.password();
-  const load = yield Api.user.initialLoad();
+  const load = yield API.user.initialLoad();
   if (!load.no_accounts) {
     if (first) {
       return new Promise((resolve, reject) => {
@@ -44,7 +44,7 @@ const getOrCreateCredentials = Co.wrap(function* (credentials, first = true) {
     throw new Error('Root credentials already created but not provided');
   }
 
-  const user = yield Api.user.create({
+  const user = yield API.user.create({
     data: { email, username, password },
     params: { d: 'undefined' }
   }, { authorize: false });
@@ -55,12 +55,12 @@ const getOrCreateCredentials = Co.wrap(function* (credentials, first = true) {
 });
 
 const getConfig = Co.wrap(function* (session) {
-  let config = yield Api.system.getConfig({ auth: session.token });
+  let config = yield API.system.getConfig({ auth: session.token });
   if (!config.TeamSettings.EnableOpenServer) {
     const data = Object.assign({}, config);
     data.TeamSettings.EnableOpenServer = true;
 
-    config = yield Api.system.updateConfig({ data }, { auth: session.token });
+    config = yield API.system.updateConfig({ data }, { auth: session.token });
   }
 
   return config;
@@ -68,7 +68,7 @@ const getConfig = Co.wrap(function* (session) {
 
 const onReady = Co.wrap(function* (callback, count = 1) {
   try {
-    yield Api.system.ping();
+    yield API.system.ping();
   } catch (err) {
     const num = count > 7 ? 1 : count;
     const timeout = Math.pow(READY_TIMEOUT, num) * 1000; // eslint-disable-line no-restricted-properties
