+++ lib/mattermost/index.js
@@ -1,11 +1,11 @@
 // Load modules
 
-const Api = require('./Api');
+const API = require('./API');
 const Service = require('./Service');
 const System = require('./System');
 
 // Deinfe exports
 
-exports.Api = Api;
+exports.API = API;
 exports.Service = Service;
 exports.System = System;
